Large Incremental GUI Additions. Project Splash Screen, Import Window.
Lots of incremental improvements and additions to the UI in this commit,
added a splash screen to allow users to select recent projects or create
new. Added more functionality to data import and integration with the
project class.
ToDo: Work on retrieving and using data from the hdf5 store created by
the project.