ENH: Improved behavior of data import code
Enhanced data import code in main.py, and the behavior relating to the
project tree view. After importing a file the flight which it was
imported to will automatically be selected as the active plot, and the
plot will be redrawn.
This also resulted in cleaning up the project.py:AirborneProject class,
removing the generate_model() method, which is now reimplemented in the
main:ProjectTreeView class, which allows for easier selecting of flights
by QModelIndex and vice versa.

Lastly, the Flight data lookup (properties) was improved by having the
Flight object cache DataFrames that had been loaded once in an instance
variable, instead of reloading from the HDFStore on every call.