BUG: Fix invalid parameter passed upon new flight creation
Line 488 in main.py was passing flight.uid to main.py::_new_plot_widget
instead of the flight object, causing a runtime error when creating a
new flight via the add_flight_dialog method.
This has been changed to pass the flight object as an attribute instead
of its uid.