ENH: Update GUI Tree View to display flight lines.
Logic added in models.py to properly handle addition of children and
sub-children to the project model.
There are still some minor issues to be worked out, i.e. properly
displaying the lines in a human-readable way.
Also still need to re-draw all lines on the plots when loading a
project.