ENH/TST: New TreeItem Class design and tests.
Added new TreeItem and interface in dgp/lib/types.py.
The Abstract base AbstractTreeItem is designed to allow polymorphic use
of classes derived from it, providing the required methods for use in a
QT TreeView or similar model.
The TreeItem implements AbstractTreeItem, providing default
implementations for the functions defined in the Abstract class.

The goal of this branch is to remove dependence on the ProjectItem class
in dgp/gui/models.py, which currently provides a wrapper for objects
added to the ProjectModel.
The idea is then that any class/object to be displayed in the GUI via a
QT model can inherit from TreeItem, and minimally implement the data() method.
A higher level of control can be implemented by the inherting class over
its display in the UI by overriding or extending methods from TreeItem.