ENH/CLN: Rewrite of ProjectModel and supporting code
ENH: This commit rewrites a large portion of the ProjectModel and TreeItem
classes, designed to improve the usability and reduce the complexity
introduced by using a wrapper class previously (ProjectItem).
The TreeItem class is designed to be inherited from for objects that
need to be displayed in a GUI TreeView.
TreeItem defines the default implementation sufficient for most objects
to immediately be useable in the ProjectModel, only the data() method
must be overriden to tell the model what to display for various roles.

TST: A basic test suite has been added to test the functionality of the
new TreeItem class. More rigorous tests should be added in the future.

CLN: Various project files have also been cleaned up to conform to a
80 character line width standard. Deprecated code and comments have also
been removed in various files.