Feature/model design (#49)
* ENH/TST: New TreeItem Class design and tests.

Added new TreeItem and interface in dgp/lib/types.py.
The Abstract base AbstractTreeItem is designed to allow polymorphic use
of classes derived from it, providing the required methods for use in a
QT TreeView or similar model.
The TreeItem implements AbstractTreeItem, providing default
implementations for the functions defined in the Abstract class.

The goal of this branch is to remove dependence on the ProjectItem class
in dgp/gui/models.py, which currently provides a wrapper for objects
added to the ProjectModel.
The idea is then that any class/object to be displayed in the GUI via a
QT model can inherit from TreeItem, and minimally implement the data() method.
A higher level of control can be implemented by the inherting class over
its display in the UI by overriding or extending methods from TreeItem.

* CLN: Reformat models/project/types modules to 80 char lines.

Reformatted project files to conform better with PEP8 80 char line
widths.
Removed old comments/general code cleanup.

* ENH/CLN: Rewrite of ProjectModel and supporting code

ENH: This commit rewrites a large portion of the ProjectModel and TreeItem
classes, designed to improve the usability and reduce the complexity
introduced by using a wrapper class previously (ProjectItem).
The TreeItem class is designed to be inherited from for objects that
need to be displayed in a GUI TreeView.
TreeItem defines the default implementation sufficient for most objects
to immediately be useable in the ProjectModel, only the data() method
must be overriden to tell the model what to display for various roles.

TST: A basic test suite has been added to test the functionality of the
new TreeItem class. More rigorous tests should be added in the future.

CLN: Various project files have also been cleaned up to conform to a
80 character line width standard. Deprecated code and comments have also
been removed in various files.

* FIX: Compatibility with Py3.5 change IntFlag to IntEnum.

* ENH/FIX: Enhanced new model code after testing.

Fixed bugs and enhanced new model code after further UI testing.
Added ability to specify styles for individual TreeItems (icon, bg/fg
color).
Enhanced update code for adding/removing/modifying items in Tree Model -
no longer need to manually specify update when adding/removing. If
modifying an attribute, a setter property should be used to call
super().update() if the attribute affects the visual appearance of an
item.
Main window UI cleanup. Adjusted margins of some widgets/panels for
cleaner look.