ENH/TST: Added datamanager module as global data interface.
Add datamanager module, to act as a global interface to
import/export/load/save data within the project.
The datamanager is initialized once and then can be used by any other
module in the project by calling datamanager.get_manager(). This
simplifies the retrieval of data from various locations in the code, as
each caller must only know the UID and type of data they need to
retrieve, as opposed to storing references to the project to utilize its
load method.

TST: Basic test suite created for datamanager - more comprehensive tests
need to be added as functionality is added to the module - plan to
support the storing of metadata and other formats such as JSON, CSV.