ENH: Various enhancements to Tab UI model.
A sprawling commit that lays the foundation for the UI as we proceed.
UI Tabs are now mostly functional, each Flight is displayed in its own
tab, and the contextual tree model (below the project tree) is
automatically switched/updated as the user selects between tabs and
sub-tabs.
This commit and the preceeding commit also introduce the datamanager
module, which was necessary at this point to reduce the complexity of
calls between the Flights/Tabs and Project. Code in various files has
been updated to utilize the new datamanager for the import/retrieval of
DataFrames.