ENH: Performance enhancement on window resize.
ENH: Added logic to lib/plotter and gui/widgets to improve GUI responsiveness when the user drag-resizes the main window, or panels within the main window.

This is accomplished by hidding the central plot widget when a
resizeEvent has been triggered, then un-hiding the widget after a short
timeout (200ms).