Rework DataSets into project replacing individual DataFiles.
Logically airborne data should be paired as a GPS DataFile and a Gravity DataFile (we'll worry about split files later), the DataSet functionality creates a group of the two distinct files, and adds an API layer to enable use of the data contained within them.
It also makes sense that Flight Lines (renamed as DataSegments) should be grouped with the particular DataSet that they are created for, as opposed to the Flight itself (which may in future enable multiple DataSets to be added).

Tests have been reorganized, added to, and split where applicable.
Various other changes were made within the project structure where needed to implement the DataSet functionality.