TEST/FIX: Testing FlightController Bindings
With the way Dialogs are currently generated for Flight context-menu
actions it is not possible to non-interactively test them, however
extracting the actions out of lambda functions within the bindings and
into their own private functions makes the intent more clear.

Also added a test to the FlightController that verifies the construction
of the _bindings list/tuple for semantic errors, i.e. checking that
QMenu declarations are valid attributes of QMenu, and that the 2-tuples
are the correct dimension.