Flatten project structure by making DataSet direct child of Flight
Flight objects can now only have DataSet's as their direct children. Sensors will be linked directly to a DataSet as opposed to linking a Sensor to a Flight object.
This is justified conceptually as a single Flight might be flown with multiple gravity meters, each collecting their own set of data (DataSet), so the sensor configuration should be linked to the data/DataSet itself, not the containing flight.