Refactor DataSet, remove unnecesarry linking to parent.
Removed path and parent params/attributes from DataSet model class. This was unnecesarry, and would complicate matters further when implementing the ability to move datasets between flights (in the future).
Also re-wrote some of the logic in the DataSet and DataSetController to enable it to lazy-load data from the HDF Store only when it was accessed for the first time.
Previously calls were being unintentionally made to load data from the DataSetController constructor, meaning that every DataFrame in the project would be loaded at project-load time - causing long/unresponsive loading.