Update project path handling, Flight clone updates.
Fix issue with FlightController clones, where clone display values were
not updated when a change was made to the original Flight/Controller.
Add parameter to AirborneProjectController to allow a project's absolute
path reference to be updated when a project is opened from a file.
Update IFlightController interface for recent API changes, removing
set_active_child method.