Provide interface to base widget for dialogs.
Add interface property (parent_widget) which returns the parent QWidget of the project model.
This is used when creating dialogs or other child GUI objects to ensure that they display with the application Icon/style if none is specified, and so that any non-modal dialogs are destroyed when the application closes.
This replaces the messy implementation previously which required the application to explicitly set the parent_widget attribtue via method call when a new project was added. The ProjectTreeModel provides this attribute as long as the model is instantiated with the parent specified.