Update parent/child interfaces.
Update and clean-up the IParent/IChild interfaces, making the interfaces more generally applicable to any controller object which is nested/has nested controllers.
Controllers can define whether they can be activated, and the interface allows for interaction via GUI input, or programatically.
The ProjectTreeModel is a special case of a Parent object, it conforms loosely to the spec for IParent, but with different names due to naming conflicts.