Add enumerations to replace string parameter matching.
Added the enumerations AxisFormatter and StateAction which define types for specifying the string formatter of a plot axis (DATETIME or SCALAR) and the StateAction enum which is used by the LineUpdate tuple to notify the interface when
a line segment has been created/updated/deleted.
Plotter backends also adds an Axis enum which simply defines the Left/Right axis pairs (for use when a plot has a twin-y axis).