Update segment label handling and decouple from parent plot object.
LinearFlightRegion now exposes signals which can be listened to by the parent to trigger an update of its siblings labels instead of directly referencing the untyped parent object.
A new signal is also added for the LFR to request a deletion (as its siblings must also be deleted by the parent [plot] object).