Transform Workspace improvements.
Added mode selection to Transform Tab - this allows the user to plot transformation results as a single line, or when toggled to break the resultant series into segments defined by the user in the line selection plot.
Add experimental method to view the source-code for the current TransformGraph function. This extracts the __init__ source code from the graph sub-class, and displays it within a QTextEdit. If the Pygments library is installed the source will also be highlighted.