Add custom DgpPlotItem with enhanced context menu.
Create DgpPlotItem as a derivative of pyqtgraph's PlotItem to override the default 'Plot Options' context menu provided by it. The default menu provides many options which are either irrelevant, or cause RuntimeErrors.
DgpPlotItem also encapsulates the process of creating a twin plot for the right y-axis, and also enables control of the twin plot (if it is enabled) via the new context menu.
Move twin y-axis plot creation into DgpPlotItem, 'right' plot can now be enabled by an init parameter.