Implement DataSegment reference tracking.
Add ability for DataSegmentController to track references to representations of itself painted in a plot. This enables actions on data-segments to be performed in a bi-directional manner, with the ability to mutate and propagate changes from both the actual plot surface, and via the DataSegmentController representation in the project tree view.
Add Delete action to DataSegmentController
Remove BaseController concrete class from DataSegmentController inheritance, use IBaseController instead.