Add integrated toolbar factory method to plots.
The base plotter class (GridPlotWidget) and its descendants now provide a get_toolbar() method which constructs a QToolBar that can be added to a UI layout.
The toolbar provides actions that are common to all plots, for example allowing the user to auto-scale the view to fit the data, hide/show grid-lines etc.
The toolbar can be extended by descendant classes to add other task specific actions, e.g. the LinePlotWidget adds a toolbar action to toggle its line selection mode.

Note that some of the icon definitions (enumerations) added here are placeholders for icons that are added in a separate future feature branch.