Implement basic state restoration for segment selection tab.
Implemented a simple state save/load mechanism utilizing QSettings
enabling the application to automatically load/plot whichever
channels were active in the plot when the tab or application was
closed.
This will pave the way for a standard interface for all workspace
tabs to implement similar persistent state tracking mechanisms.