Implement improved observer pattern/registration.
Improved the observer registration method to enable passing of
bound methods, and observation of specific state actions (update/delete)

This allows UI objects (tabs/widgets etc) to explicitly register and
subscribe to controller events - e.g. when a controller object is
deleted, it will execute a callback notifying all observers of its
destruction.

Added clone tracking feature to AbstractController to standardize
the update of clones, and integrate the state notifications with those
of the observers.