Annotate coverage skip on GUI action methods.
Many GUI actions/methods to activate them are not easily tested,
and would only be minorly useful to have tested anyways.
pragma: no cover annotations have been added and will be added in
future to these sorts of blocks/methods as I believe it is a waste
of time to attempt to test many of them.